Init commit: HW 2 Q3
The task for Q3 of HW2 is to complete main.cpp where any input from any of the included .txt files will return a vector of average values given the filter width and the data. The filter width dictates how many values are included in the average calculation and what the middle number (where the data centers upon) is located. If a data point within the initial vector cannot be centered or included in the filter width, it returns it's initial value.

Length/size of the vector should be the same as the initial data given that we are calculating the averages for the filter width and those that aren't included will return as initial values.

Values returned are fixed to 4 decimal places and outputted to a .txt file titled OutFilexx.txt. In the output file, each element of the filtered vector is separated by a space and comma, excluding the last data point.

Signed-off-by: CieranW <cieranwong@gmail.com>