Began writing the Filter function to go through the dataset and depending on the status (valid or not?) assigns it to a temp vector to perform an avg calculation and get all of the valid data in one spot.
Still not sure how I want to do the filtering, currently have it in a separate vector, but I'm thinking that we might just keep the original vector and just use if-else statements to verify if the status of the data is valid. This'll allow for easier access to modifying the data when changing the data status from valid to filtered but it may prove complicated when trying to perform calculations. Perhaps I may include additional if-else statements or a separate vector just to perform calculations before inserting into the original vector.

Will come back to it later on.

Signed-off-by: CieranW <cieranwong@gmail.com>