All done!
Got 100/100 on the tests so I'm good to go now. Had to revamp the entire Filter() function after I realised that a valid filter window included the other statuses. So if you had a data set that went: valid, valid, undefined, valid, valid ... those first two valid points weren't included in the filter window. Additionally, had to shift around some of the classification requirements in Classify() so that they would not be reclassified by another test. Added a couple of comments here and there explaining the code and left main.cpp in how I would run on my computer.

Signed-off-by: CieranW <cieranwong@gmail.com>